app/requirements.txt

arlo==1.2.59
requests==2.27.1
PyYAML==6.0
todoist-api-python==1.1.1
